gatchat: Allow configuring a custom header length
Some Motorola modems use a non-standard U1234 style packet ID in front
of all the messages. As the data is already coming in TS 27.010 packets,
we have currently no use for this custom packet header.

Let's add basic handling for configuring a custom header and skipping
it. More handling can be added as needed.

Note that also a custom gatsyntax handler is likely needed for modems
using a custom packet header.